+++ backend/docker/sql/backend-prd.sql
@@ -213,6 +213,18 @@ CREATE TABLE IF NOT EXISTS edp(
   PRIMARY KEY ( key )
 );
 
+CREATE TABLE IF NOT EXISTS faucet(
+  id serial NOT NULL,
+  sender TEXT,
+  value TEXT,
+  txnHash TEXT,
+  destination TEXT,
+  network TEXT,
+  address INET,
+  createdAt timestamp default current_timestamp,
+  PRIMARY KEY( id )
+);
+
 INSERT INTO total (name, count) VALUES
   ('blocks', 0),
   ('extrinsics', 0),
@@ -257,6 +269,7 @@ GRANT ALL PRIVILEGES ON TABLE featured TO prd_cerestats;
 GRANT ALL PRIVILEGES ON TABLE account TO prd_cerestats;
 GRANT ALL PRIVILEGES ON TABLE total TO prd_cerestats;
 GRANT ALL PRIVILEGES ON TABLE edp TO prd_cerestats;
+GRANT ALL PRIVILEGES ON TABLE faucet TO prd_cerestats;
 
 --
 -- Fast counters